+++ xclim/core/bootstrapping.py
@@ -7,7 +7,8 @@ import xarray
 from boltons.funcutils import wraps
 from xarray.core.dataarray import DataArray
 
-from xclim.core.calendar import convert_calendar, parse_offset, percentile_doy
+from .calendar import parse_offset, percentile_doy
+from .utils import convert_calendar
 
 
 def percentile_bootstrap(func):
