+++ xclim/testing/tests/test_indicators.py
@@ -22,9 +22,8 @@ from xclim.core.formatting import (
 )
 from xclim.core.indicator import Daily, Indicator, ResamplingIndicator, registry
 from xclim.core.units import convert_units_to, declare_units, units
-from xclim.core.utils import InputKind, MissingVariableError
+from xclim.core.utils import InputKind, MissingVariableError, select_time
 from xclim.indices import tg_mean
-from xclim.indices.generic import select_time
 from xclim.testing import open_dataset
